https://github.com/GreenInfo-Network/east-coast-greenway/issues/119 refactor GPX export, add button to mobile cue sheet
